requirements.txt
fastapi
uvicorn
pydantic[email]
pydantic-settings
sqlalchemy
duckdb
duckdb-engine
python-multipart
pandas
yfinance

python-jose[cryptography]
bcrypt
httpx
requests
google-genai
google-api-core
apscheduler
python-dateutil
xlrd>=2.0.1
openpyxl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow
PyMuPDF
